refactor : 패키지 구조 변경 및 Exception에 메시지 추가
