--- src/main/java/Step1/Calculator.java
@@ -1,20 +1,0 @@
-package Step1;
-
-
-public class Calculator {
-    public int addition(int num1, int num2) {
-        return num1 + num2;
-    }
-
-    public int subtraction(int num1, int num2) {
-        return num1 - num2;
-    }
-
-    public int multiplication(int num1, int num2) {
-        return num1 * num2;
-    }
-
-    public int division(int num1, int num2) {
-        return num1 / num2;
-    }
-}
