+++ src/test/java/stringCalculator/MathEquationValidatorTest.java
@@ -1,9 +1,10 @@
-package Step1;
+package stringCalculator;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
+import stringCalculator.validator.MathEquationValidator;
 
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
